Split PassphraseModal into logic and presentation components
